--- lib/wake/azure/exec.rb
@@ -1,14 +1,0 @@
-require "wake/powershell"
-
-module Wake
-  module_function
-
-  def exec(command)
-    if Wake.powershell?
-      system command
-      exit $?.exitstatus
-    else
-      Kernel.exec command
-    end
-  end
-end
